+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 ### Changed
 - Update dependency `seriaport` to v4.0
+- **Breaking change**: Simpler Error type
 
 ## v0.3.1 (2020-08-15)
 - Fix documentation